Fix api để chạy frontend

+++ src/main/java/com/example/mini_project/controllers/UserController.java
@@ -2,6 +2,7 @@ package com.example.mini_project.controllers;
 import com.example.mini_project.entities.ResponseDTO;
 import com.example.mini_project.entities.User;
 import com.example.mini_project.repositories.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.http.HttpStatus;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.*;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Slf4j
 @RestController
@@ -40,7 +42,7 @@ public class UserController {
 
     @GetMapping("/search")
     public ResponseEntity<ResponseDTO<List<User>>> searchPeople(
-            @RequestParam(name="username", required = false) String username,
+//            @RequestParam(name="username", required = false) String username,
             @RequestParam(name="firstname", required = false) String firstname,
             @RequestParam(name="lastname", required = false) String lastname,
             @RequestParam(name="phoneNum", required = false) String phoneNum,
@@ -54,25 +56,29 @@ public class UserController {
             begin = LocalDate.now().minusYears(age + 1).plusDays(1);
             end = LocalDate.now().minusYears(age);
         }
-        List<User> results = userRepository.searchByCustom(username, firstname, lastname, address, phoneNum, cityname, begin, end);
+        List<User> results = userRepository.searchByCustom(firstname, lastname, address, phoneNum, cityname, begin, end);
         return buildResponse(HttpStatus.OK, "Tìm kiếm thành công!", results);
     }
 
     @PreAuthorize("hasRole('ADMIN')")
-    @Cacheable(value="users", key="#id")
-    @GetMapping("/{id}")
-    public ResponseEntity<ResponseDTO<User>> getUserById(@PathVariable("id") Long id) {
-        return userRepository.findById(id)
+    @GetMapping("/{username}")
+    public ResponseEntity<ResponseDTO<User>> getByUsername(@PathVariable("username") String username) {
+        return userRepository.findByUsername(username)
                 .map(user -> buildResponse(HttpStatus.OK, "User fetched successfully", user))
                 .orElseGet(() -> buildResponse(HttpStatus.NOT_FOUND, "User not found", null));
     }
 
     @PreAuthorize("hasRole('ADMIN')")
-    @PostMapping("/")
-    public ResponseEntity<ResponseDTO<User>> createNewUser(@RequestBody @Validated User user, BindingResult bindingResult) throws BindException{
+    @PostMapping("/create")
+    public ResponseEntity<ResponseDTO<User>> createNewUser(@RequestBody @Validated User user, BindingResult bindingResult){
         if (bindingResult.hasErrors()) {
-            throw new BindException(bindingResult);
-        }
+            String message = bindingResult.getFieldErrors()
+                    .stream()
+                    .map(e -> e.getField() + ": " + e.getDefaultMessage())
+                    .collect(Collectors.joining("; "));
+
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(new ResponseDTO<>(message, null));        }
         user.setPassword(encoder.encode(user.getPassword()));
         User saved = userRepository.save(user);
         return buildResponse(HttpStatus.CREATED, "User created successfully", saved);
@@ -81,13 +87,15 @@ public class UserController {
     /**
      * Chỉ thay đổi được nếu là admin or là chủ sở hữu thông tin
      */
-    @PutMapping("/{id}")
-    public ResponseEntity<ResponseDTO<User>> updateUser(@PathVariable("id") Long id, @RequestBody User user,
-                                                        Authentication auth) {
-        if (!isOwnerOrAdmin(auth, id)) {
+    @PutMapping("/update/{username}")
+    public ResponseEntity<ResponseDTO<User>> updateUser(@PathVariable("username") String username, @RequestBody User user,
+                                                        Authentication auth, HttpServletRequest request) {
+        log.info("Received request URI: {}", request.getRequestURI());
+        log.info("Path variable username: {}", username);
+        if (!isOwnerOrAdmin(auth, username)) {
             return buildResponse(HttpStatus.BAD_REQUEST, "Bạn chỉ có thể edit thông tin của bạn!", null);
         }
-        Optional<User> userOptional = userRepository.findById(id);
+        Optional<User> userOptional = userRepository.findByUsername(username);
         if (userOptional.isEmpty()) {
             return buildResponse(HttpStatus.BAD_REQUEST, "Không có gì để update", null);
         }
@@ -95,77 +103,74 @@ public class UserController {
         if (user.getPassword() != null) {
             updatedUser.setPassword(encoder.encode(user.getPassword()));
         }
-        if (!user.getAuthorities().isEmpty()) {
-            user.getAuthorities().forEach(a -> updatedUser.addRole(a.getAuthority()));
-        }
         userRepository.save(updatedUser);
         return buildResponse(HttpStatus.OK, "Thay đổi thông tin thành công!", updatedUser);
     }
     // Helper method
     private User getUser(User user, Optional<User> userOptional) {
         User updatedUser = userOptional.get();
-        if (user.getUsername() != null) {
-            updatedUser.setUsername(user.getUsername());
-        }
-        if (user.getAddress() != null) {
+        if (user.getAddress() != null && !user.getAddress().isEmpty()) {
             updatedUser.setAddress(user.getAddress());
         }
         if (user.getDob() != null) {
             updatedUser.setDob(user.getDob());
         }
-        if (user.getFirstname() != null) {
+        if (user.getFirstname() != null && !user.getFirstname().isEmpty()) {
             updatedUser.setFirstname(user.getFirstname());
         }
-        if (user.getLastname() != null) {
+        if (user.getLastname() != null && !user.getLastname().isEmpty()) {
             updatedUser.setLastname(user.getLastname());
         }
-        if (user.getPhoneNum() != null) {
+        if (user.getPhoneNum() != null && !user.getPhoneNum().isEmpty()) {
             updatedUser.setPhoneNum(user.getPhoneNum());
         }
         return updatedUser;
     }
-    // Helper method
-    private boolean isOwnerOrAdmin(Authentication auth, Long currentID) {
+
+    /**
+     * Helper method, check principal của authentication với 1 entry fetch từ db theo username.
+     * Note: auth la mot object duoc tao ra tu token?
+     * @param auth
+     * @param updateUsername
+     * @return
+     */
+    private boolean isOwnerOrAdmin(Authentication auth, String updateUsername) {
         boolean isAdmin =  auth.getAuthorities().stream()
                 .anyMatch(a->a.getAuthority().equals("ROLE_ADMIN"));
-        Optional<User> userOptional = userRepository.findById(currentID);
-        // Nếu có store userID ở trong authentication thì sẽ hiệu quả hơn.
-        if (userOptional.isEmpty()) return isAdmin;
-        User user = userOptional.get();
+        if (isAdmin) return true;
         Object principal = auth.getPrincipal();
+
         boolean isOwner = false;
         if (principal instanceof UserDetails userDetails) {
-            isOwner = user.getUsername().equals(userDetails.getUsername());
+            isOwner = updateUsername.equals(userDetails.getUsername());
         }
         if (principal instanceof String p) {
-            isOwner = user.getUsername().equals(p);
+            isOwner = updateUsername.equals(p);
         }
-        if (isOwner) log.info("User là chủ sở hữu");
-        else log.info("User không phải chủ sở hữu");
-        return isAdmin || isOwner;
+        if (isOwner) log.info("User là chủ sở hữu"); // debug
+        else log.info("User không phải chủ sở hữu"); // debug
+        return isOwner;
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/delete/{username}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<ResponseDTO<User>> deleteUser(@PathVariable("id") Long id, Authentication auth){
-        Optional<User> optionalUser = userRepository.findById(id);
+    public ResponseEntity<ResponseDTO<User>> deleteUser(@PathVariable("username") String username, Authentication auth){
+        Optional<User> optionalUser = userRepository.findByUsername(username);
         if (optionalUser.isEmpty()) {
-            return buildResponse(HttpStatus.BAD_REQUEST, "User Id không tồn tại để xóa", null);
+            return buildResponse(HttpStatus.BAD_REQUEST, "User không tồn tại để xóa", null);
         }
+
         User user = optionalUser.get();
+        boolean containAdmin = user.getAuthorities().stream()
+                .anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"));
+
+        if (containAdmin) {
+            return buildResponse(HttpStatus.BAD_REQUEST, "Không thể xóa admin role", null);
+        }
         userRepository.delete(user);
         return buildResponse(HttpStatus.OK, "Xóa thành công user", user);
     }
 
-    @ExceptionHandler(BindException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public String handleBindException(BindException e) {
-        List<String> errors = e.getFieldErrors()
-                .stream()
-                .map(error -> error.getField() + ": " + error.getDefaultMessage())
-                .toList();
-        throw new IllegalArgumentException("Vi phạm yêu cầu của Database tại: "  + errors);
-    }
 
     private <T> ResponseEntity<ResponseDTO<T>> buildResponse(HttpStatus status, String message, T data) {
         return ResponseEntity.status(status).body(new ResponseDTO<>(message, data));
